refactor(go-mod): Refactored GoModDetectableOptions to GoModCliDetectableOptions.
