+++ detectable/src/main/java/com/synopsys/integration/detectable/detectables/go/gomod/GoModCliDetectable.java
@@ -16,7 +16,6 @@ import com.synopsys.integration.detectable.detectable.annotation.DetectableInfo;
 import com.synopsys.integration.detectable.detectable.exception.DetectableException;
 import com.synopsys.integration.detectable.detectable.executable.resolver.GoResolver;
 import com.synopsys.integration.detectable.detectable.result.DetectableResult;
-import com.synopsys.integration.detectable.detectables.go.GoModDetectableOptions;
 import com.synopsys.integration.detectable.extraction.Extraction;
 import com.synopsys.integration.detectable.extraction.ExtractionEnvironment;
 
@@ -27,16 +26,16 @@ public class GoModCliDetectable extends Detectable {
     private final FileFinder fileFinder;
     private final GoResolver goResolver;
     private final GoModCliExtractor goModCliExtractor;
-    private final GoModDetectableOptions goModDetectableOptions;
+    private final GoModCliDetectableOptions goModCliDetectableOptions;
 
     private ExecutableTarget goExe;
 
-    public GoModCliDetectable(DetectableEnvironment environment, FileFinder fileFinder, GoResolver goResolver, GoModCliExtractor goModCliExtractor, GoModDetectableOptions goModDetectableOptions) {
+    public GoModCliDetectable(DetectableEnvironment environment, FileFinder fileFinder, GoResolver goResolver, GoModCliExtractor goModCliExtractor, GoModCliDetectableOptions goModCliDetectableOptions) {
         super(environment);
         this.fileFinder = fileFinder;
         this.goResolver = goResolver;
         this.goModCliExtractor = goModCliExtractor;
-        this.goModDetectableOptions = goModDetectableOptions;
+        this.goModCliDetectableOptions = goModCliDetectableOptions;
     }
 
     @Override
@@ -55,6 +54,6 @@ public class GoModCliDetectable extends Detectable {
 
     @Override
     public Extraction extract(ExtractionEnvironment extractionEnvironment) {
-        return goModCliExtractor.extract(environment.getDirectory(), goExe, goModDetectableOptions.isDependencyVerificationEnabled());
+        return goModCliExtractor.extract(environment.getDirectory(), goExe, goModCliDetectableOptions.isDependencyVerificationEnabled());
     }
 }
